moving gesamtspielplan to own go file
